+++ src/action/createSessionStartedUserAction.ts
@@ -1,12 +1,12 @@
 import viewport from '../utils/viewport'
 import location from '../utils/location'
-import { EventType, GravitySessionStartedEvent } from '../types'
+import { UserActionType, SessionStartedUserAction } from '../types'
 import { config } from '../config'
 import gravityDocument from '../utils/gravityDocument'
 
-export function createSessionStartedEvent(): GravitySessionStartedEvent {
-  const initSessionEvent: GravitySessionStartedEvent = {
-    type: EventType.SessionStarted,
+export function createSessionStartedUserAction(): SessionStartedUserAction {
+  const action: SessionStartedUserAction = {
+    type: UserActionType.SessionStarted,
     recordedAt: new Date().toISOString(),
     location: location(),
     document: gravityDocument(),
@@ -17,8 +17,8 @@ export function createSessionStartedEvent(): GravitySessionStartedEvent {
 
   const cypress = (window as any).Cypress
   if (cypress?.currentTest !== undefined) {
-    initSessionEvent.test = cypress.currentTest.titlePath.join(' > ')
+    action.test = cypress.currentTest.titlePath.join(' > ')
   }
 
-  return initSessionEvent
+  return action
 }