+++ src/collector/CollectorWrapper.ts
@@ -1,21 +1,21 @@
 import { v4 as uuidv4 } from 'uuid'
 import ClickEventListener from '../event/listener/ClickEventListener'
-import { createSessionStartedEvent } from '../event/createSessionStartedEvent'
+import { createSessionStartedUserAction } from '../action/createSessionStartedUserAction'
 import { CollectorOptions } from '../types'
 import ChangeEventListener from '../event/listener/ChangeEventListener'
 import EventHandler from '../event/handler/EventHandler'
 import UnloadEventListener from '../event/listener/UnloadEventListener'
-import { debugEventSessionSender, defaultEventSessionSender } from '../event/handler/eventSessionSender'
 import KeyUpEventListener from '../event/listener/KeyUpEventListener'
 import KeyDownEventListener from '../event/listener/KeyDownEventListener'
+import { debugUserActionSessionSender, defaultUserActionSessionSender } from '../event/handler/userActionSessionSender'
 
 class CollectorWrapper {
   readonly eventHandler: EventHandler
 
   constructor(options: CollectorOptions, private readonly window: Window) {
     const output = options.debug
-      ? debugEventSessionSender(options.maxDelay)
-      : defaultEventSessionSender(options.authKey, options.gravityServerUrl)
+      ? debugUserActionSessionSender(options.maxDelay)
+      : defaultUserActionSessionSender(options.authKey, options.gravityServerUrl)
     this.eventHandler = new EventHandler(uuidv4(), options.requestInterval, output)
 
     this.initSession()
@@ -23,7 +23,7 @@ class CollectorWrapper {
   }
 
   private initSession() {
-    return this.eventHandler.run(createSessionStartedEvent())
+    return this.eventHandler.run(createSessionStartedUserAction())
   }
 
   private initializeEventHandlers() {