+++ src/event/listener/ClickEventListener.ts
@@ -1,17 +1,17 @@
 import EventHandler from '../handler/EventHandler'
-import { createGravityEvent } from '../createGravityEvent'
+import { createTargetedUserAction } from '../../action/createTargetedUserAction'
 import EventListener from './EventListener'
-import { EventType } from '../../types'
+import { UserActionType } from '../../types'
 
 class ClickEventListener extends EventListener {
   constructor(eventHandler: EventHandler, window: Window) {
-    super(eventHandler, EventType.Click, window)
+    super(eventHandler, UserActionType.Click, window)
   }
 
   listener(event: MouseEvent) {
     const pointerEvent = event as PointerEvent
     if (pointerEvent.pointerType !== undefined && pointerEvent.pointerType !== '') {
-      this.eventHandler.run(createGravityEvent(event, this.eventType))
+      this.eventHandler.run(createTargetedUserAction(event, this.userActionType))
     }
   }
 }