+++ src/event/listener/EventListener.ts
@@ -1,15 +1,15 @@
 import EventHandler from '../handler/EventHandler'
-import { EventType } from '../../types'
+import { UserActionType } from '../../types'
 
 export default abstract class EventListener {
   protected constructor(
     protected readonly eventHandler: EventHandler,
-    protected readonly eventType: EventType,
+    protected readonly userActionType: UserActionType,
     protected readonly window: Window,
   ) {}
 
   init(): void {
-    this.window.addEventListener(this.eventType, this.listener.bind(this), true)
+    this.window.addEventListener(this.userActionType, this.listener.bind(this), true)
   }
 
   protected abstract listener(event: Event): void