+++ src/event/listener/KeyDownEventListener.ts
@@ -1,18 +1,22 @@
 import EventHandler from '../handler/EventHandler'
-import { createGravityEvent } from '../createGravityEvent'
+import { createTargetedUserAction } from '../../action/createTargetedUserAction'
 import EventListener from './EventListener'
-import { EventType } from '../../types'
+import { UserActionType } from '../../types'
 import { isKeyAllowedByKeyListeners, isTargetAllowedByKeyListeners } from '../../utils/listeners'
 
 class KeyDownEventListener extends EventListener {
   constructor(eventHandler: EventHandler, window: Window) {
-    super(eventHandler, EventType.KeyDown, window)
+    super(eventHandler, UserActionType.KeyDown, window)
   }
 
   listener(event: KeyboardEvent) {
-    if (isKeyAllowedByKeyListeners(event.code)) return this.eventHandler.run(createGravityEvent(event, this.eventType))
+    if (isKeyAllowedByKeyListeners(event.code)) {
+      return this.eventHandler.run(createTargetedUserAction(event, this.userActionType))
+    }
 
-    if (isTargetAllowedByKeyListeners(event.target)) this.eventHandler.run(createGravityEvent(event, this.eventType))
+    if (isTargetAllowedByKeyListeners(event.target)) {
+      this.eventHandler.run(createTargetedUserAction(event, this.userActionType))
+    }
   }
 }