+++ src/utils/completeOptions.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest'
-import { GRAVITY_SERVER_ADDRESS } from '../event/handler/eventSessionSender'
 import { CollectorOptions } from '../types'
 import completeOptions from './completeOptions'
+import { GRAVITY_SERVER_ADDRESS } from '../event/handler/userActionSessionSender'
 
 describe('completeOptions', () => {
   it('throws an error when not provided a CollectorOptions', () => {